+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+## [0.1.7] - 2026-07-18
+
+## Other
+
+- update on 2026.07.18 16:03:50 (Release 0.1.5) (Release 0.1.5)
+- update on 2026.07.17 16:13:19
+- andiamo avanti... (Release 0.1.5) (Release 0.1.5)
+- andiamo avanti... (Release 0.1.5) (Release 0.1.5)
+- sostituito  Color dal logger con colors.get_colors() (Release 0.1.5) (Release 0.1.5)
+- avanzamento (Release 0.1.5) (Release 0.1.5)
+- modificato gitStatus (Release 0.1.5) (Release 0.1.5)
+- update on 2026.07.11 17:29:44 (Release 0.1.5) (Release 0.1.5)
+
+## [0.1.6] - 2026-07-18
+
+## Other
+
+- update on 2026.07.18 16:03:50 (Release 0.1.5) (Release 0.1.5)
+- update on 2026.07.17 16:13:19
+- andiamo avanti... (Release 0.1.5) (Release 0.1.5)
+- andiamo avanti... (Release 0.1.5) (Release 0.1.5)
+- sostituito  Color dal logger con colors.get_colors() (Release 0.1.5) (Release 0.1.5)
+- avanzamento (Release 0.1.5) (Release 0.1.5)
+- modificato gitStatus (Release 0.1.5) (Release 0.1.5)
+- update on 2026.07.11 17:29:44 (Release 0.1.5) (Release 0.1.5)
+
+## [0.1.6] - 2026-07-18
+
+## Other
+
+- update on 2026.07.18 16:03:50 (Release 0.1.5) (Release 0.1.5)
+- update on 2026.07.17 16:13:19
+- andiamo avanti... (Release 0.1.5) (Release 0.1.5)
+- andiamo avanti... (Release 0.1.5) (Release 0.1.5)
+- sostituito  Color dal logger con colors.get_colors() (Release 0.1.5) (Release 0.1.5)
+- avanzamento (Release 0.1.5) (Release 0.1.5)
+- modificato gitStatus (Release 0.1.5) (Release 0.1.5)
+- update on 2026.07.11 17:29:44 (Release 0.1.5) (Release 0.1.5)
+
 ## v0.1.4 - 2026-06-09
 
 ### Fixes
@@ -18,4 +57,3 @@
 - update on 2026.05.22 18:28:49 (Release 0.0.1)
 - initial commit (Release 0.0.1)
 - Initial commit
-